fix: repair Rust workspace build, lint and tests (pyo3 0.29)
- Remove invalid generate-abi3 pyo3 feature (build was broken)
- Fix clippy -D warnings across all crates: redundant closures,
  needless_range_loop, manual_next_back, len_zero, manual range contains,
  unused imports, deprecated #[pyclass] FromPyObject (from_py_object)
- Wire cryptobot_py submodules: register features/risk/orderbook/backtest
  pymodules; PyModule::new API; make #[pymodule] fns pub
- Correct Kelly test expectation (f* = 0.4 for p=0.6, b=2)

cargo fmt --check, clippy -D warnings, cargo test --workspace all green (31 tests)

+++ crates/cryptobot-core/src/lib.rs
@@ -15,6 +15,7 @@ pub fn placeholder() -> &'static str {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use rust_decimal::Decimal;
 
     #[test]
     fn placeholder_returns_name() {
@@ -24,7 +25,7 @@ mod tests {
     #[test]
     fn test_ohlcv_bar_creation() {
         let bar = OhlcvBar::new(50000.0, 50100.0, 49900.0, 50050.0, 1000.0);
-        assert_eq!(bar.open, 50000.0);
+        assert_eq!(bar.open, Decimal::from(50000));
         assert!(bar.is_valid());
     }
